feat: PersistentManager 클래스 생성
기본 기능 구현 및 그에 따른 로직 수정
