+++ Diary/Application/AppDelegate.swift
@@ -27,30 +27,4 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
       sessionRole: connectingSceneSession.role
     )
   }
-
-  // MARK: - Core Data stack
-
-  lazy var persistentContainer: NSPersistentContainer = {
-    let container = NSPersistentContainer(name: "Diary")
-    container.loadPersistentStores(completionHandler: { (_, error) in
-      if let error = error as NSError? {
-        fatalError("Unresolved error \(error), \(error.userInfo)")
-      }
-    })
-    return container
-  }()
-
-  // MARK: - Core Data Saving support
-
-  func saveContext () {
-    let context = persistentContainer.viewContext
-    if context.hasChanges {
-      do {
-        try context.save()
-      } catch {
-        let nserror = error as NSError
-        fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
-      }
-    }
-  }
 }
