+++ Diary/Application/SceneDelegate.swift
@@ -23,6 +23,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
   }
 
   func sceneDidEnterBackground(_ scene: UIScene) {
-    (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+    PersistentManager.shared.saveContext()
   }
 }
